+++ src/main/kotlin/model/BiomePaths.kt
@@ -88,8 +88,8 @@ data class BiomePaths(
 
                         points.add(
                             Point(
-                                x = pairSplit[0].toInt().toShort(),
-                                y = pairSplit[1].toInt().toShort()
+                                x = pairSplit[0].toInt(),
+                                y = pairSplit[1].toInt()
                             )
                         )
                     }
